// src/BindingChaos.Web/src/api/models/UserGroupListItemResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BindingChaos.Web.Gateway
 * BindingChaos Web Gateway API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Represents a user group item in a list response.
 * @export
 * @interface UserGroupListItemResponse
 */
export interface UserGroupListItemResponse {
    /**
     * The unique identifier of the user group.
     * @type {string}
     * @memberof UserGroupListItemResponse
     */
    id?: string;
    /**
     * The identifier of the commons this group governs.
     * @type {string}
     * @memberof UserGroupListItemResponse
     */
    commonsId?: string;
    /**
     * The name of the user group.
     * @type {string}
     * @memberof UserGroupListItemResponse
     */
    name?: string;
    /**
     * The philosophy of the user group.
     * @type {string}
     * @memberof UserGroupListItemResponse
     */
    philosophy?: string;
    /**
     * The pseudonym of the participant who founded the group.
     * @type {string}
     * @memberof UserGroupListItemResponse
     */
    foundedByPseudonym?: string;
    /**
     * The timestamp when the group was formed.
     * @type {Date}
     * @memberof UserGroupListItemResponse
     */
    formedAt?: Date;
    /**
     * The current number of members in the group.
     * @type {number}
     * @memberof UserGroupListItemResponse
     */
    memberCount?: number;
    /**
     * The join policy of the group.
     * @type {string}
     * @memberof UserGroupListItemResponse
     */
    joinPolicy?: string;
}

/**
 * Check if a given object implements the UserGroupListItemResponse interface.
 */
export function instanceOfUserGroupListItemResponse(value: object): value is UserGroupListItemResponse {
    return true;
}

export function UserGroupListItemResponseFromJSON(json: any): UserGroupListItemResponse {
    return UserGroupListItemResponseFromJSONTyped(json, false);
}

export function UserGroupListItemResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserGroupListItemResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'commonsId': json['commonsId'] == null ? undefined : json['commonsId'],
        'name': json['name'] == null ? undefined : json['name'],
        'philosophy': json['philosophy'] == null ? undefined : json['philosophy'],
        'foundedByPseudonym': json['foundedByPseudonym'] == null ? undefined : json['foundedByPseudonym'],
        'formedAt': json['formedAt'] == null ? undefined : (new Date(json['formedAt'])),
        'memberCount': json['memberCount'] == null ? undefined : json['memberCount'],
        'joinPolicy': json['joinPolicy'] == null ? undefined : json['joinPolicy'],
    };
}

export function UserGroupListItemResponseToJSON(json: any): UserGroupListItemResponse {
    return UserGroupListItemResponseToJSONTyped(json, false);
}

export function UserGroupListItemResponseToJSONTyped(value?: UserGroupListItemResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'commonsId': value['commonsId'],
        'name': value['name'],
        'philosophy': value['philosophy'],
        'foundedByPseudonym': value['foundedByPseudonym'],
        'formedAt': value['formedAt'] == null ? undefined : ((value['formedAt']).toISOString()),
        'memberCount': value['memberCount'],
        'joinPolicy': value['joinPolicy'],
    };
}